+++ opensearchapi/api.delete_by_query.go
@@ -315,7 +315,7 @@ func (r DeleteByQueryRequest) Do(ctx context.Context, transport Transport) (*Res
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.