+++ opensearchapi/api.delete_script.go
@@ -156,7 +156,7 @@ func (r DeleteScriptRequest) Do(ctx context.Context, transport Transport) (*Resp
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.