+++ opensearchapi/api.indices.create.go
@@ -166,7 +166,7 @@ func (r IndicesCreateRequest) Do(ctx context.Context, transport Transport) (*Res
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.