+++ opensearchapi/api.indices.exists_index_template.go
@@ -162,7 +162,7 @@ func (r IndicesExistsIndexTemplateRequest) Do(ctx context.Context, transport Tra
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.