+++ opensearchapi/api.indices.get_settings.go
@@ -190,7 +190,7 @@ func (r IndicesGetSettingsRequest) Do(ctx context.Context, transport Transport) 
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.