+++ opensearchapi/api.mget.go
@@ -190,7 +190,7 @@ func (r MgetRequest) Do(ctx context.Context, transport Transport) (*Response, er
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.