+++ opensearchapi/api.ping.go
@@ -134,7 +134,7 @@ func (r PingRequest) Do(ctx context.Context, transport Transport) (*Response, er
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.