+++ opensearchapi/api.snapshot.get.go
@@ -175,7 +175,7 @@ func (r SnapshotGetRequest) Do(ctx context.Context, transport Transport) (*Respo
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.