+++ opensearchapi/api.tasks.cancel.go
@@ -167,7 +167,7 @@ func (r TasksCancelRequest) Do(ctx context.Context, transport Transport) (*Respo
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.