+++ opensearchapi/api.termvectors.go
@@ -208,7 +208,7 @@ func (r TermvectorsRequest) Do(ctx context.Context, transport Transport) (*Respo
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.