+++ opensearchapi/api.update.go
@@ -212,7 +212,7 @@ func (r UpdateRequest) Do(ctx context.Context, transport Transport) (*Response, 
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.