+++ opensearchapi/api.update_by_query_rethrottle.go
@@ -148,7 +148,7 @@ func (r UpdateByQueryRethrottleRequest) Do(ctx context.Context, transport Transp
 		Header:     res.Header,
 	}
 
-	return &response, nil
+	return &response, response.Err()
 }
 
 // WithContext sets the request context.